a better trajectory generator is implemented.

+++ src/main/java/indexingTopology/util/taxi/Car.java
@@ -7,7 +7,7 @@ public class Car {
     public long id;
     public double x;
     public double y;
-    public Car(int id, double x, double y) {
+    public Car(long id, double x, double y) {
         this.id = id;
         this.x = x;
         this.y = y;
